Add Rakefile for generating corpus for search
This currently creates a `corpus.json` from all the guides on the site.
We will iterate over this and use it as an index for searching the
Playbook.

There's going to be an added challenge when we want to search in
`index.markdown` as it's not particularly structured but that can come
later.

The approach here was taken from [this helpful blog
post](https://zoeleblanc.com/blog/adding-search-to-jekyll-sites-with-lunr/),
with a couple of small tweaks.

We'll probably need to find a way of removing extra `\n`s which are
output to the json file due to the way we wrap lines in Markdown files
as this is going to look weird in search results and possibly make
searching over newlines impossible.